[REFACT] Query용과 Command용으로 컨트롤러 및 서비스 분리
